moved templates to .h and fixed overloading problem in CommunitySize
