+++ DataFormats/HGCalReco/interface/TICLGraph.h
@@ -46,7 +46,9 @@ public:
   }*/
 };
 
-bool operator==(ElementaryNode const& eN1, ElementaryNode const& eN2);
+inline bool operator==(ElementaryNode const& eN1, ElementaryNode const& eN2) {
+  return (eN1.getId() == eN2.getId()) && (eN1.getNeighbours() == eN2.getNeighbours());
+}
 
 // a node can consist of one or more elementary nodes (needed to implement the aggregate graph)
 template <class T>
@@ -67,6 +69,11 @@ public:
     return ((internalStructure_[0]).size() == 0) ? true : false;
   }
 };
+
+template <class T>
+bool operator==(Node<T> const& n1, Node<T> const& n2) {
+  return ((n1.getInternalStructure()) == (n2.getInternalStructure()));
+}
 
 //tested this implementation on godbolt
 template <class T>
@@ -143,6 +150,99 @@ public:
   }*/
 };
 
+//fills an empty vector with the result of flattening operation of a community
+template <class T>
+std::vector<ElementaryNode>& flatCommunity(std::vector<Node<T>> const& community,
+                                           std::vector<ElementaryNode>& flattenedCommunity) {
+  assert(flattenedCommunity.empty());
+  for (auto const& node : community) {
+    if (node.isNodeDegreeZero()) {
+      for (auto const& elementaryNode : node.getInternalStructure())
+        flattenedCommunity.push_back(elementaryNode);
+    } else
+      flatCommunity(node.getInternalStructure(), flattenedCommunity);
+  }
+  return flattenedCommunity;
+}
+
+// the number of edges b/w 2 nodes is the number of edges between their elementary nodes
+template <class T>
+int numberOfEdges(std::vector<Node<T>> const& communityA, std::vector<Node<T>> const& communityB) {
+  int numberOfEdges{};
+
+  std::vector<ElementaryNode> flattenedCommunityA{};
+  std::vector<ElementaryNode> flattenedCommunityB{};
+  flatCommunity(communityA, flattenedCommunityA);
+  flatCommunity(communityB, flattenedCommunityB);
+
+  for (auto const& elementaryNodeA : flattenedCommunityA) {
+    std::vector<unsigned int> const& neighboursA{elementaryNodeA.getNeighbours()};
+    for (auto const& Id : neighboursA) {
+      auto it{std::find_if(flattenedCommunityB.begin(), flattenedCommunityB.end(), [=](ElementaryNode const& elNodeB) {
+        return (elNodeB.getId()) == Id;
+      })};
+      if (it != flattenedCommunityB.end()) {
+        ++numberOfEdges;
+      }
+    }
+  }
+  assert(numberOfEdges >= 0);
+  return numberOfEdges;
+}
+
+inline int communitySize(std::vector<ElementaryNode> const& community, int size) { return size; }
+
+//the size of a community is the number of elementary nodes in it
+template <class T>
+int communitySize(std::vector<Node<T>> const& community, int size = 0) {
+  for (auto const& node : community) {
+    if (node.isNodeDegreeZero()) {
+      size += (node.getInternalStructure()).size();
+    } else
+      size = communitySize(node.getInternalStructure(), size);
+  }
+  return size;
+}
+
+// two nodes are neighbours if at least two of their elementary nodes are
+template <class T>
+bool areNeighbours(Node<T> const& nodeA, Node<T> const& nodeB) {
+  std::vector<Node<T>> A{nodeA};
+  std::vector<Node<T>> B{nodeB};
+  std::vector<ElementaryNode> flattenedCommunityA{};
+  std::vector<ElementaryNode> flattenedCommunityB{};
+  flatCommunity(A, flattenedCommunityA);
+  flatCommunity(B, flattenedCommunityB);
+  bool result{false};
+  for (auto const& elementaryNodeA : flattenedCommunityA) {
+    std::vector<unsigned int> const& neighboursA{elementaryNodeA.getNeighbours()};
+    for (auto const& Id : neighboursA) {
+      auto it{std::find_if(flattenedCommunityB.begin(), flattenedCommunityB.end(), [=](ElementaryNode const& elNodeB) {
+        return (elNodeB.getId()) == Id;
+      })};
+      if (it != flattenedCommunityB.end()) {
+        result = true;
+        break;
+      }
+    }
+  }
+  return result;
+}
+
+//tells me if community is contained within a certain subset
+template <class T>
+bool isCommunityContained(std::vector<Node<T>> const& community, std::vector<Node<T>> const& subset) {
+  bool isContained{true};
+  for (auto const& node : community) {
+    auto it{std::find(subset.begin(), subset.end(), node)};
+    if (it == subset.end()) {
+      isContained = false;
+      break;
+    }
+  }
+  return isContained;
+}
+
 //takes a community and return the vector of all the Elementary Nodes in the community
 template <class T>
 std::vector<ElementaryNode>& flatCommunity(std::vector<Node<T>> const& community,
@@ -187,17 +287,5 @@ public:
     }
   }
 };
-
-template <class T>
-int numberOfEdges(std::vector<Node<T>> const& communityA, std::vector<Node<T>> const& communityB);
-
-template <class T>
-int communitySize(std::vector<Node<T>> const& community, int size = 0);
-
-template <class T>
-bool areNeighbours(Node<T> const& nodeA, Node<T> const& nodeB);
-
-template <class T>
-bool isCommunityContained(std::vector<Node<T>> const& community, std::vector<Node<T>> const& subset);
 
 #endif
